+++ source/Tests/presetManager.py
@@ -2,34 +2,8 @@ import os
 import shutil
 import subprocess
 import json
-import sys
-import stringProcessor
-import win32com.client # pip install pywin32
-
-class Preset:
-    def __init__(self, name:str, description:str, registryLocation:str, files:list):
-        self.name = name
-        self.description = description
-        self.registryLocation = registryLocation
-        self.files = files
-
-    def toJson(self):
-        '''
-        Returns the object as a json string
-        '''
-        return json.dumps(self, default=lambda o: o.__dict__)
-
-    def to_dict(self):
-        '''
-        Returns the object as a dictionary
-        '''
-        return {"name": self.name, "description": self.description, "registryLocation": self.registryLocation, "files": self.files}
-    
-    def to_Preset(dict:dict):
-        """
-        returns a Preset object with the name,description and registryLocation values
-        """
-        return Preset(dict['name'],dict['description'],dict['registryLocation'],dict['files'])
+from preset_class import Preset
+from shortcut_util import shortcututil
 
 class presetmanager:
 
@@ -37,13 +11,11 @@ class presetmanager:
     global repositoryDirectory
     global presetListFilename
     global homeDirectory
-    global powerShellExePath
 
     homeDirectory = "C:\\Users\\" + os.getenv("username") 
     presetsDirectory = homeDirectory + "\\AppData\\Local\\DLO\\Presets"
     repositoryDirectory = homeDirectory + "\\AppData\\Local\\DLO\\Presets\\Repository"
     presetListFilename = homeDirectory + "\\AppData\\Local\\DLO\\Presets\\PresetList.json"
-    powerShellExePath = "C:\\Windows\\System32\\WindowsPowerShell\\v1.0\\powershell.exe"
 
 
     @staticmethod
@@ -94,12 +66,11 @@ class presetmanager:
 
             if os.path.splitext(filepath)[1] == ".lnk":
 
-                target_path = return_shortcut_target(filepath)
+                target_path = shortcututil.return_shortcut_target(filepath)
                 files.append({"path": target_path, "name": os.path.basename(target_path)})
             else: 
-                
-                filepath = shutil.copy(filepath,os.path.join(repositoryDirectory))
-                os.remove(filepath)
+                # filepath = shutil.copy(filepath,os.path.join(repositoryDirectory))
+                # os.remove(filepath)
                 files.append({"path": filepath, "name": os.path.basename(filepath)})
 
 
@@ -266,18 +237,6 @@ class presetmanager:
         :return: True if the file is empty, False otherwise\n
         '''
         return os.stat(filename).st_size == 0
-    
-@staticmethod
-def return_shortcut_target(shortcut_path:str):
-    '''
-    Returns the target of a shortcut (.lnk)\n
-    :param shortcut_path: The path to the shortcut\n
-    :return: The target of the shortcut\n
-    '''
-
-    shell = win32com.client.Dispatch("WScript.Shell")
-    shortcut = shell.CreateShortCut(shortcut_path)
-    return shortcut.Targetpath
 
 @staticmethod
 def getPresetsInJsonFormat(presetsRaw:list):
@@ -289,46 +248,3 @@ def getPresetsInJsonFormat(presetsRaw:list):
     presets.sort(key=lambda obj: obj["name"])
     presetsJSON = json.dumps({"presets": presets})
     return presetsJSON
-
-@staticmethod
-def createShortcut(sourcePath:str,destinationFolder:str,ps1FilePath:str):
-    """
-    Creates a shortcut for each file in the sourcePath and saves it in the destinationFolder
-    destinationFolder: Folder where the shortcuts should be saved
-    ps1FilePath: Path to the PowerShell Script that creates the shortcuts
-    sourcePath: Folder where the files are located
-    """
-    sourcePath = stringProcessor.process(sourcePath)
-    ps1FilePath = stringProcessor.process(ps1FilePath)
-
-    try:
-        #all entries of the targetFolder
-        entries = os.listdir(sourcePath)
-    except:
-        print("-> File can not be found")
-        exit()
-
-    if "desktop.ini" in entries:
-        ## we don't want to work with the file "desktop.ini" therefore i'm deleting it from the list
-        entries.remove("desktop.ini")
-
-    for entry in entries:
-
-        ## Parameters being sent to the PowerShell Script to create a shortcut
-        targetFile = os.path.join(sourcePath,entry)
-        shortcutPath = os.path.join(destinationFolder,entry)
-        
-        ## DEBUG
-        ## print ( " -> Printing variables used for PS execution..." )
-        ## print ( "powerShellExePath: " + powerShellExePath )
-        ## print ( "ps1FilePath: " + ps1FilePath )
-        ## print ( "shortcutpath: " + shortcutPath )
-        ## print ( "targetFile: " + targetFile)
-
-        p = subprocess.run([powerShellExePath,ps1FilePath,shortcutPath,targetFile],stdout=sys.stdout)
-
-        if p.returncode == 0:
-            print("-> Succesfully created shortcut for '" + targetFile +"'")
-        else:
-            print("-> An error as has occured creating the Shortcut")
-            exit()
