+++ source/eelui/presetManager.py
@@ -2,47 +2,24 @@ import os
 import shutil
 import subprocess
 import json
-import win32com.client # pip install pywin32
-import stringProcessor
-import sys
-
-class Preset:
-    def __init__(self, name, description, registryLocation, files):
-        self.name = name
-        self.description = description
-        self.registryLocation = registryLocation
-        self.files = files
-
-    def toJson(self):
-        return json.dumps(self, default=lambda o: o.__dict__)
-
-    def to_dict(self):
-        return {"name": self.name, "description": self.description, "registryLocation": self.registryLocation, "files": self.files}
-    
-    def to_Preset(dict):
-        """dict as a parameter
-         \n returns a Preset object with the name,description and registryLocation values
-        """
-        return Preset(dict['name'],dict['description'],dict['registryLocation'],dict['files'])
+from preset_class import Preset
+from shortcut_util import shortcututil
 
 class presetmanager:
 
-    global developermode
-    developermode=True
     global presetsDirectory
     global repositoryDirectory
     global presetListFilename
     global homeDirectory
-    global powerShellExePath
-    
+
     homeDirectory = "C:\\Users\\" + os.getenv("username") 
     presetsDirectory = homeDirectory + "\\AppData\\Local\\DLO\\Presets"
-    repositoryDirectory = homeDirectory + "\\AppData\\Local\\DLO\\Repository"
+    repositoryDirectory = homeDirectory + "\\AppData\\Local\\DLO\\Presets\\Repository"
     presetListFilename = homeDirectory + "\\AppData\\Local\\DLO\\Presets\\PresetList.json"
-    powerShellExePath = "C:\\Windows\\System32\\WindowsPowerShell\\v1.0\\powershell.exe"
+
 
     @staticmethod
-    def create_preset(presetName, presetDescription):
+    def create_preset(presetName:str, presetDescription:str):
         '''
         Creates a new preset with given presetName and Description\n
         presetName: Name of the preset\n
@@ -53,8 +30,6 @@ class presetmanager:
 
         # Set the directory where the files are located
         desktop_path = os.path.join(os.path.expanduser('~'), 'Desktop') # Getting the desktop path
-        if developermode==True:
-            desktop_path = os.path.join(desktop_path, 'DLO-Virtual-Desktop') # Getting the desktop path
         presetsRaw = []
 
         # 1: Export the registry key to the registryDirectory
@@ -83,27 +58,28 @@ class presetmanager:
         # Getting the files from the desktop
         file_list = [f for f in os.listdir(desktop_path) ]
         file_paths = [os.path.join(desktop_path, f) for f in file_list]
-
-        # Save files in current preset in as objects of the form {path: 'abc', name: 'xyz'} (example)
+            
+        #Adding files to the preset
         files = [] 
         for filepath in file_paths:
+            #copy files to repository and then delete them from desktop
+
             if os.path.splitext(filepath)[1] == ".lnk":
-                target_path = return_shortcut_target(filepath)
+
+                target_path = shortcututil.return_shortcut_target(filepath)
                 files.append({"path": target_path, "name": os.path.basename(target_path)})
             else: 
-                # The following code was moved to another method named 'copy_file_to_repository':
-                ######################################################################
                 # filepath = shutil.copy(filepath,os.path.join(repositoryDirectory))
                 # os.remove(filepath)
-                ######################################################################
                 files.append({"path": filepath, "name": os.path.basename(filepath)})
+
 
         # Adding the new entry to the 'presetlist.json'
         presetsRaw.append(
             Preset(f"{presetName}",
             f"{presetDescription}",
             f"C:\\Users\\Bence\\Desktop\\{presetName}.reg",
-            file_paths)) 
+            files)) 
 
         # Converting Objects into a dictionary
         presetsJSON = getPresetsInJsonFormat(presetsRaw)
@@ -111,22 +87,14 @@ class presetmanager:
         with open(f"{presetListFilename}", "w") as outfile:
             outfile.write(presetsJSON)
 
-        return file_paths
-
     @staticmethod
-    def copy_file_to_repository(filepath):
-        print("recieved filepath: " + filepath)
-        dest=shutil.copy(filepath, repositoryDirectory)
-        destdir, filename = os.path.split(filepath)
-        print("> Creating shortcut... Data:")
-        print(">    Source: " + dest)
-        print(">    Destination: " + destdir)
-        # createShortcut(sourcePath:str,destinationFolder:str,ps1FilePath:str)
-        createShortcut(dest, destdir, "C:\\Users\\Bence\\OneDrive - HTBLA Leonding\\Schule\\SYP - Systemplanung\\Desktop-Layout-Organizer\\source\\createShortcutCommands.ps1")
-        os.remove(filepath)
-
-    @staticmethod
-    def change_preset(presetName, presetNewName, presetNewDesc):
+    def change_preset(presetName:str, presetNewName:str, presetNewDesc:str):
+        '''
+        Changes the name and description of a preset\n
+        presetName: Name of the preset\n
+        presetNewName: New name of the preset\n
+        presetNewDesc: New description of the preset
+        '''
         presetmanager.create_directories()
 
         # Change to the registry directory
@@ -151,7 +119,11 @@ class presetmanager:
                         outfile.write(presetsJSON)
 
     @staticmethod
-    def save_preset(presetName):
+    def save_preset(presetName:str):
+        '''
+        Saves the current desktop layout as a preset\n
+        presetName: Name of the preset
+        '''
         presetmanager.create_directories()
         # Delete the existing preset file
         os.remove(os.path.join(presetsDirectory, presetName + ".reg"))
@@ -159,7 +131,11 @@ class presetmanager:
         subprocess.call(f"reg export HKEY_CURRENT_USER\\SOFTWARE\\Microsoft\\Windows\\Shell\\Bags\\1\\Desktop {os.path.join(presetsDirectory, presetName)}.reg", shell=True)
 
     @staticmethod
-    def load_preset(presetName):
+    def load_preset(presetName:str):
+        '''
+        Loads a preset onto the Desktop\n
+        presetName: Name of the preset
+        '''
         presetmanager.create_directories()
         # Getting name of registry file
         presetFileName = os.path.join(presetsDirectory, presetName + ".reg")
@@ -174,7 +150,11 @@ class presetmanager:
         subprocess.Popen(['explorer'], shell=True, stdin=None, stdout=None, stderr=None, close_fds=True)
 
     @staticmethod
-    def delete_preset(presetToDelete):
+    def delete_preset(presetToDelete:str):
+        '''
+        Deletes a preset\n
+        presetToDelete: Name of the preset
+        '''
         presetmanager.create_directories()
 
         # Change to the registry directory
@@ -211,6 +191,9 @@ class presetmanager:
 
     @staticmethod
     def get_all_entries():
+        '''
+        Returns a list of all presets
+        '''
         presetmanager.create_directories()
         
         f = open(f"{presetListFilename}", "r")
@@ -225,6 +208,9 @@ class presetmanager:
 
     @staticmethod
     def create_directories():
+        '''
+        Creates the necessary directories if they don't exist
+        '''
         
         # Check if the directories exist
         if not os.path.exists(presetsDirectory) and os.path.exists(repositoryDirectory):
@@ -244,65 +230,21 @@ class presetmanager:
                 outfile.write(presetsJSON)
   
     @staticmethod
-    def fileIsEmpty(filename):
+    def fileIsEmpty(filename:str):
         '''
         Checks if a file is empty by confirming that its size is 0 bytes\n
         :param filename: The file to check\n
         :return: True if the file is empty, False otherwise\n
         '''
         return os.stat(filename).st_size == 0
-    
-@staticmethod
-def return_shortcut_target(shortcut_path):
-    '''
-    Returns the target of a shortcut (.lnk)\n
-    :param shortcut_path: The path to the shortcut\n
-    :return: The target of the shortcut\n
-    '''
-
-    shell = win32com.client.Dispatch("WScript.Shell")
-    shortcut = shell.CreateShortCut(shortcut_path)
-    return shortcut.Targetpath
 
 @staticmethod
-def getPresetsInJsonFormat(presetsRaw):
+def getPresetsInJsonFormat(presetsRaw:list):
+    '''
+    Returns the presets in a JSON format\n
+    presetsRaw: List of Preset objects\n
+    '''
     presets = [obj.to_dict() for obj in presetsRaw]
     presets.sort(key=lambda obj: obj["name"])
     presetsJSON = json.dumps({"presets": presets})
     return presetsJSON
-
-@staticmethod
-def createShortcut(fileToCopy:str,destinationFolder:str,ps1FilePath:str):
-    """
-    Creates a shortcut for each file in the sourcePath and saves it in the destinationFolder
-    destinationFolder: Folder where the shortcuts should be saved
-    ps1FilePath: Path to the PowerShell Script that creates the shortcuts
-    sourcePath: Folder where the files are located
-    """
-    sourcePath = stringProcessor.process(fileToCopy)
-    ps1FilePath = stringProcessor.process(ps1FilePath)
-
-    shortcutPath = "undefined"
-
-    # Set the directory where the files are located
-    desktop_path = os.path.join(os.path.expanduser('~'), 'Desktop') # Getting the desktop path
-    if developermode==True:
-        desktop_path = os.path.join(desktop_path, 'DLO-Virtual-Desktop') # Getting the desktop path
-
-    shortcutPath = os.path.join(desktop_path, os.path.splitext(os.path.basename(fileToCopy))[0] + ".lnk")
-    print("> shortcutpath: " +shortcutPath)
-    print("filetocopy: " + fileToCopy)
-    
-    ## DEBUG
-    ## print ( " -> Printing variables used for PS execution..." )
-    ## print ( "powerShellExePath: " + powerShellExePath )
-    ## print ( "ps1FilePath: " + ps1FilePath )
-    ## print ( "shortcutpath: " + shortcutPath )
-    ## print ( "targetFile: " + targetFile)
-    
-    p = subprocess.run([powerShellExePath,ps1FilePath, shortcutPath, fileToCopy],stdout=sys.stdout)
-
-    if p.returncode == 0:
-        print("-> Succesfully created shortcut for '" + fileToCopy +"'")
-    else:
-        print("-> An error as has occured creating the Shortcut")
