+++ source/eelui/requirements.txt
@@ -1,1 +1,2 @@
 eel
+pywin32
